+++ day2/ex1/defs/Decoder.hpp
@@ -10,8 +10,8 @@ class Decoder
         std::ifstream file;
         std::string input;
         std::vector<std::string> ranges;
-        int password = 0;
-        std::vector<int> idLimits;
+        long password = 0;
+        std::vector<long> idLimits;
     public:
         void addNewFile(const char *rute);
         void parseInput();
@@ -20,9 +20,9 @@ class Decoder
         void idAnalyzer(std::string id);
         void parseLimits(const std::string range);
 
-        int getPassword() const;
+        long getPassword() const;
         std::vector<std::string> getRanges() const;
-        std::vector<int> getLimits() const;
+        std::vector<long> getLimits() const;
 
         void clearLimits();
 };
